requirements.txt
discord.py>=2.3.0
aiofiles>=23.0.0